refactor(clappr-core): drop jest-directory-named-resolver
The resolver was the last package-local Jest dependency and the only
custom resolution left in the monorepo. It is unmaintained (it still
declares jest ^19) and it never calls options.defaultResolver, going
straight to the resolve package instead, so it ignores exports maps and
conditions that Jest 28+ relies on.

Only test files depended on the directory-named lookup: src/ already
imports the full path to the module file. Spell out that path in the 43
affected test imports and drop the resolver, so clappr-core runs on the
stock Jest resolver like every other package.

Suite unchanged: 28 suites, 433 tests, coverage 77.93/70.35/66.43/78.69.
