+++ packages/clappr-core/package.json
@@ -43,7 +43,6 @@
     "@rollup/plugin-alias": "^5.0.0",
     "@rollup/plugin-json": "^6.0.0",
     "@rollup/plugin-replace": "^5.0.2",
-    "jest-directory-named-resolver": "^0.3.0",
     "rollup-plugin-filesize": "^10.0.0",
     "rollup-plugin-serve": "^2.0.2",
     "typescript": "^5.5.2"
