+++ packages/clappr-core/src/base/ui_container_plugin/ui_container_plugin.test.js
@@ -1,5 +1,5 @@
 import UIContainerPlugin from './ui_container_plugin'
-import ErrorMixin from '@/base/error_mixin'
+import ErrorMixin from '@/base/error_mixin/error_mixin'
 
 describe('UI Container Plugin', () => {
   describe('constructor', () => {
